Add files via upload
Attached files are sample code from a software development class. In this class I gained experience building class diagrams, using these diagrams to form classes in java, and also creating interactive webpages built from jsp.

This project code will not be able to be ran as we used data from an SQL Database in which I no longer have access to. However I hope you will be able to gain a viewpoint in my capabilities. 

The project gave different styles of accesss to users and admins, depending  on their login credentials. Some data was shared such as viewing a university, but other things were different such as editing a university for admins, and adding a university to a saved school list for regular users. We used a mixture of java code as well as jsp code to limit what was available for the person logging in.